gio: manually implement content_type_guess
