+++ gio/Gir.toml
@@ -249,9 +249,9 @@ status = "generate"
     ignore = true
     [[object.function]]
     name = "content_type_guess"
-        [[object.function.parameter]]
-        name = "filename"
-        string_type = "filename"
+    # implemented manually until gir generates nullable array parameters
+    # https://github.com/gtk-rs/gir/issues/1133
+    manual = true
 
 [[object]]
 name = "Gio.ActionGroup"
